+++ src/effects/stop.js
@@ -1,15 +1,17 @@
 import { _ } from '../utils/underscore.js';
 import { Events } from '../events.js';
+import { Element } from '../element.js';
 
 /**
  * @name Two.Stop
  * @class
+ * @extends Two.Element
  * @param {Number} [offset] - The offset percentage of the stop represented as a zero-to-one value. Default value flip flops from zero-to-one as new stops are created.
  * @param {String} [color] - The color of the stop. Default value flip flops from white to black as new stops are created.
  * @param {Number} [opacity] - The opacity value. Default value is 1, cannot be lower than 0.
  * @nota-bene Used specifically in conjunction with {@link Two.Gradient}s to control color graduation.
  */
-export class Stop extends Events {
+export class Stop extends Element {
 
   constructor(offset, color, opacity) {
 
@@ -19,13 +21,6 @@ export class Stop extends Events {
       Object.defineProperty(this, prop, proto[prop]);
     }
 
-    /**
-     * @name Two.Stop#renderer
-     * @property {Object}
-     * @description Object access to store relevant renderer specific variables. Warning: manipulating this object can create unintended consequences.
-     * @nota-bene With the {@link Two.SvgRenderer} you can access the underlying SVG element created via `shape.renderer.elem`.
-     */
-    this.renderer = {};
     this._renderer.type = 'stop';
 
     /**
@@ -155,15 +150,6 @@ const proto = {
       if (this.parent) {
         this.parent._flagStops = true;
       }
-    }
-  },
-  renderer: {
-    enumerable: false,
-    get: function() {
-      return this._renderer;
-    },
-    set: function(obj) {
-      this._renderer = obj;
     }
   }
 };
